requirements.txt
yfinance>=0.2.36
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0.0
streamlit>=1.35.0
plotly>=5.20.0
reportlab>=4.0.0
